--- a/synfig-core/src/synfig/general.h
+++ b/synfig-core/src/synfig/general.h
@@ -88,6 +88,15 @@ extern void warning(const String &str);
 extern void info(const char *format,...);
 extern void info(const String &str);
 
+//#define LOGGING_ENABLED
+#ifdef LOGGING_ENABLED
+#define DEBUG_LOG(logger, fmt, ...) if (getenv(logger)) printf(fmt __VA_OPT__(,) __VA_ARGS__);
+#define DEBUG_GETENV(name) getenv(name)
+#else
+#define DEBUG_LOG(logger, fmt, ...) void (0)
+#define DEBUG_GETENV(name) false
+#endif
+
 //! Returns absolute path to the binary
 extern String get_binary_path(const String &fallback_path);