--- a/synfig-core/src/synfig/layers/layer_skeleton.cpp
+++ b/synfig-core/src/synfig/layers/layer_skeleton.cpp
@@ -33,6 +33,9 @@
 #endif
 
 #include "layer_skeleton.h"
+#ifdef _DEBUG
+#include "synfig/general.h"
+#endif
 
 #include <synfig/localization.h>
 
@@ -90,8 +93,8 @@ Layer_Skeleton::Layer_Skeleton():
 #ifdef _DEBUG
 Layer_Skeleton::~Layer_Skeleton()
 {
-	if (getenv("SYNFIG_DEBUG_DESTRUCTORS"))
-		printf("%s:%d ~Layer_Skeleton()\n", __FILE__, __LINE__);
+	 DEBUG_LOG("SYNFIG_DEBUG_DESTRUCTORS",
+		"%s:%d ~Layer_Skeleton()\n", __FILE__, __LINE__);
 }
 #endif