… synfig-core/src/synfig/valuenode.cpp
… synfig-core/src/synfig/valuenodes/valuenode_average.cpp
… synfig-core/src/synfig/valuenodes/valuenode_bline.cpp
… synfig-core/src/synfig/valuenodes/valuenode_blinecalctangent.cpp
… synfig-core/src/synfig/valuenodes/valuenode_blinecalcwidth.cpp
… synfig-core/src/synfig/valuenodes/valuenode_blinereversetangent.cpp
… synfig-core/src/synfig/valuenodes/valuenode_bone.cpp
… synfig-core/src/synfig/valuenodes/valuenode_boneinfluence.cpp
… synfig-core/src/synfig/valuenodes/valuenode_bonelink.cpp
… synfig-core/src/synfig/valuenodes/valuenode_boneweightpair.cpp
… synfig-core/src/synfig/valuenodes/valuenode_compare.cpp
… synfig-core/src/synfig/valuenodes/valuenode_composite.cpp
… synfig-core/src/synfig/valuenodes/valuenode_const.cpp
… synfig-core/src/synfig/valuenodes/valuenode_cos.cpp
… synfig-core/src/synfig/valuenodes/valuenode_derivative.cpp
… synfig-core/src/synfig/valuenodes/valuenode_dilist.cpp
… synfig-core/src/synfig/valuenodes/valuenode_dotproduct.cpp
… synfig-core/src/synfig/valuenodes/valuenode_duplicate.cpp
… synfig-core/src/synfig/valuenodes/valuenode_dynamic.cpp
… synfig-core/src/synfig/valuenodes/valuenode_dynamiclist.cpp
… synfig-core/src/synfig/valuenodes/valuenode_exp.cpp
… synfig-core/src/synfig/valuenodes/valuenode_gradientcolor.cpp
… synfig-core/src/synfig/valuenodes/valuenode_gradientrotate.cpp
… synfig-core/src/synfig/valuenodes/valuenode_integer.cpp
… synfig-core/src/synfig/valuenodes/valuenode_intstring.cpp
… synfig-core/src/synfig/valuenodes/valuenode_join.cpp
… synfig-core/src/synfig/valuenodes/valuenode_linear.cpp
… synfig-core/src/synfig/valuenodes/valuenode_log.cpp
… synfig-core/src/synfig/valuenodes/valuenode_modulo.cpp
… synfig-core/src/synfig/valuenodes/valuenode_not.cpp